✨ feat: Update settings.json and config.yaml
This commit updates the `.vscode/settings.json` file, adding the "businesscard" and "netboot" options to the list of supported features. It also modifies the `os/config.yaml` file, changing the value of `lb.debianInstaller` to "cdrom" and `lb.debianInstallerDistribution` to "bookworm". These changes improve the configuration and compatibility of the system.